Rename the drawer folder to drawers
