+++ src/drawers/__mocks__/vertex-drawer-mock.ts
