+++ src/drawers/vertex-drawer.spec.ts
