+++ src/drawers/vertex-drawer.ts
