+++ src/tools/arrow-edit-tool.ts
@@ -1,9 +1,9 @@
 import {Tool} from "./tool";
 import {ToolService} from "./tool-service";
 import {EntitySelectionService} from "./entity-selection-service";
-import {VertexDrawer} from "../drawer/vertex-drawer";
+import {VertexDrawer} from "../drawers/vertex-drawer";
 import {ArrowEntity} from "../entities/arrow-entity";
-import {Vertex} from "../drawer/vertex";
+import {Vertex} from "../drawers/vertex";
 import {ArrowVertexFactory} from "./arrow-vertex-factory";
 import {ArrowModifyType} from "./arrow-modify-tool";
 import {LayerService} from "../layer-service";
