+++ src/tools/box-edit-tool.ts
@@ -1,7 +1,7 @@
 import {Tool} from "./tool";
-import {SelectBoxDrawer} from "../drawer/select-box-drawer";
+import {SelectBoxDrawer} from "../drawers/select-box-drawer";
 import {BoxEntity} from "../entities/box-entity";
-import {SelectBox} from "../drawer/select-box";
+import {SelectBox} from "../drawers/select-box";
 import {ResizeType} from "./box-resize-tool";
 import {ToolService} from "./tool-service";
 import {EntitySelectionService} from "./entity-selection-service";
