+++ src/tools/text-create-tool.ts
@@ -3,10 +3,10 @@ import {LayerService} from "../layer-service";
 import {Entity} from "../entities/entity";
 import {TextEntity} from "../entities/text-entity";
 import {EntityIdService} from "../entities/entity-id-service";
-import {Text} from "../drawer/text";
-import {TextDrawer} from "../drawer/text-drawer";
-import {CursorDrawer} from "../drawer/cursor-drawer";
-import {Cursor} from "../drawer/cursor";
+import {Text} from "../drawers/text";
+import {TextDrawer} from "../drawers/text-drawer";
+import {CursorDrawer} from "../drawers/cursor-drawer";
+import {Cursor} from "../drawers/cursor";
 
 export class TextCreateTool implements Tool {
