+++ src/tools/text-move-tool.ts
@@ -1,14 +1,14 @@
 import {Tool} from "./tool";
 import {LayerService} from "../layer-service";
-import {SelectBoxDrawer} from "../drawer/select-box-drawer";
+import {SelectBoxDrawer} from "../drawers/select-box-drawer";
 import Constants from "../constants";
-import {BoxDrawer} from "../drawer/box-drawer";
+import {BoxDrawer} from "../drawers/box-drawer";
 import {ToolService} from "./tool-service";
 import {TextEntity} from "../entities/text-entity";
-import {Vertex} from "../drawer/vertex";
-import {VertexDrawer} from "../drawer/vertex-drawer";
-import {Text} from "../drawer/text";
-import {TextDrawer} from "../drawer/text-drawer";
+import {Vertex} from "../drawers/vertex";
+import {VertexDrawer} from "../drawers/vertex-drawer";
+import {Text} from "../drawers/text";
+import {TextDrawer} from "../drawers/text-drawer";
 
 export class TextMoveTool implements Tool {
