+++ src/tools/tool-service.ts
@@ -3,9 +3,9 @@ import {ArrowCreateTool} from "./arrow-create-tool";
 import {BoxCreateTool} from "./box-create-tool";
 import {TextCreateTool} from "./text-create-tool";
 import {LayerService} from "../layer-service";
-import {SelectBoxDrawer} from "../drawer/select-box-drawer";
+import {SelectBoxDrawer} from "../drawers/select-box-drawer";
 import {SelectTool} from "./select-tool";
-import {BoxDrawer} from "../drawer/box-drawer";
+import {BoxDrawer} from "../drawers/box-drawer";
 import {EntityIdService} from "../entities/entity-id-service";
 import {EntitySelectionService} from "./entity-selection-service";
 import {BoxEditTool} from "./box-edit-tool";
@@ -13,12 +13,12 @@ import {BoxEntity} from "../entities/box-entity";
 import {TextEntity} from "../entities/text-entity";
 import {TextEditTool} from "./text-edit-tool";
 import {BoxResizeTool, ResizeType} from "./box-resize-tool";
-import {TextDrawer} from "../drawer/text-drawer";
-import {CursorDrawer} from "../drawer/cursor-drawer";
+import {TextDrawer} from "../drawers/text-drawer";
+import {CursorDrawer} from "../drawers/cursor-drawer";
 import {BoxMoveTool} from "./box-move-tool";
-import {VertexDrawer} from "../drawer/vertex-drawer";
+import {VertexDrawer} from "../drawers/vertex-drawer";
 import {TextMoveTool} from "./text-move-tool";
-import {ArrowDrawer} from "../drawer/arrow-drawer";
+import {ArrowDrawer} from "../drawers/arrow-drawer";
 import {ArrowEditTool} from "./arrow-edit-tool";
 import {ArrowEntity} from "../entities/arrow-entity";
 import {ArrowFlipTool} from "./arrow-flip-tool";
